requirements.txt
azure-functions
requests==2.22.0
numpy==1.17.4
Pillow==6.2.1
azure-storage-blob==12.1.0
azure-keyvault==4.0.0
azure-cosmos==3.1.2
azure-mgmt-resource==6.0.0
msrestazure==0.6.2